Created MainScreen + separate drawer
